feat(gains-network): Add locked gTokens positions on Arbitrum
