+++ src/apps/gains-network/gains-network.module.ts
@@ -3,6 +3,7 @@ import { Module } from '@nestjs/common';
 import { AbstractApp } from '~app/app.dynamic-module';
 
 import { ArbitrumGainsNetworkGTokenTokenFetcher } from './arbitrum/gains-network.g-token.token-fetcher';
+import { ArbitrumGainsNetworkLockedContractPositionFetcher } from './arbitrum/gains-network.locked.contract-position-fetcher';
 import { ArbitrumGainsNetworkStakingContractPositionFetcher } from './arbitrum/gains-network.staking.contract-position-fetcher';
 import { GainsNetworkContractFactory } from './contracts';
 import { PolygonGainsNetworkGTokenTokenFetcher } from './polygon/gains-network.g-token.token-fetcher';
@@ -12,11 +13,14 @@ import { PolygonGainsNetworkStakingContractPositionFetcher } from './polygon/gai
 @Module({
   providers: [
     GainsNetworkContractFactory,
+    // Arbitrum
+    ArbitrumGainsNetworkGTokenTokenFetcher,
+    ArbitrumGainsNetworkStakingContractPositionFetcher,
+    ArbitrumGainsNetworkLockedContractPositionFetcher,
+    // Polygon
     PolygonGainsNetworkStakingContractPositionFetcher,
     PolygonGainsNetworkGTokenTokenFetcher,
     PolygonGainsNetworkLockedContractPositionFetcher,
-    ArbitrumGainsNetworkGTokenTokenFetcher,
-    ArbitrumGainsNetworkStakingContractPositionFetcher,
   ],
 })
 export class GainsNetworkAppModule extends AbstractApp() {}
